requirements.txt
flask
python-dotenv
numpy
pandas
scikit-learn
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
groq
matplotlib